feat: change /user to /auth route, bb-20
