+++ apps/mobile/src/libs/packages/storage/libs/enums/storage-key.enum.ts
@@ -1,5 +1,6 @@
 const StorageKey = {
 	TOKEN: "token",
+	USER: "user",
 } as const;
 
 export { StorageKey };
