+++ apps/mobile/src/slices/auth/actions.ts
@@ -1,12 +1,27 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
+import { storage, StorageKey } from "~/libs/packages/storage/storage";
 import { type AsyncThunkConfig } from "~/libs/types/types";
 import {
+	type UserGetAllItemResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "~/packages/users/users";
 
 import { name as sliceName } from "./auth.slice";
+
+const getAuthenticatedUser = createAsyncThunk<
+	UserGetAllItemResponseDto,
+	undefined,
+	AsyncThunkConfig
+>(`${sliceName}/get-authenticated-user`, async (_, { extra }) => {
+	const { authApi } = extra;
+	const response: UserGetAllItemResponseDto =
+		await authApi.getAuthenticatedUser();
+	await storage.set(StorageKey.USER, JSON.stringify(response));
+
+	return response;
+});
 
 const signUp = createAsyncThunk<
 	UserSignUpResponseDto,
@@ -18,4 +33,4 @@ const signUp = createAsyncThunk<
 	return authApi.signUp(signUpPayload);
 });
 
-export { signUp };
+export { getAuthenticatedUser, signUp };
