+++ apps/mobile/src/slices/auth/auth.ts
@@ -1,8 +1,9 @@
-import { signUp } from "./actions";
+import { getAuthenticatedUser, signUp } from "./actions";
 import { actions } from "./auth.slice";
 
 const allActions = {
 	...actions,
+	getAuthenticatedUser,
 	signUp,
 };
