+++ apps/mobile/src/slices/users/actions.ts
@@ -1,10 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types";
-import {
-	type UserGetAllItemResponseDto,
-	type UserGetAllResponseDto,
-} from "~/packages/users/users";
+import { type UserGetAllResponseDto } from "~/packages/users/users";
 
 import { name as sliceName } from "./users.slice";
 
@@ -18,14 +15,4 @@ const loadAll = createAsyncThunk<
 	return userApi.getAll();
 });
 
-const getAuthenticatedUser = createAsyncThunk<
-	UserGetAllItemResponseDto,
-	number,
-	AsyncThunkConfig
->(`${sliceName}/get-authenticated-user`, async (userId, { extra }) => {
-	const { userApi } = extra;
-
-	return await userApi.getAuthenticatedUser(userId);
-});
-
-export { getAuthenticatedUser, loadAll };
+export { loadAll };
